backend/requirements.txt
# Core FastAPI
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Database & ORM
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.9

# Data validation and settings
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0

# Authentication & Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Environment & Configuration
python-dotenv==1.1.1

# Security enhancements
slowapi==0.1.9  # Rate limiting
secure==1.0.1   # Security headers

# Performance & Monitoring
redis==6.2.0    # Caching
prometheus-client==0.22.1  # Metrics

# Additional dependencies (auto-installed with above packages)
# starlette==0.47.1
# pydantic-core==2.33.2
# anyio==4.9.0
# click==8.2.1
# h11==0.16.0
# httptools==0.6.4
# watchfiles==1.1.0
# websockets==15.0.1
# bcrypt==4.3.0
# dnspython==2.7.0
# limits==5.4.0 